Trim whitespace before parsing response string
This fixes a bug where the response string could have whitespace/newline characters are the end, we had it occur on the bp10emu endpoint.  The problem with that is when you do `$bluePay->isSuccessfulResponse()` it will return `false` if the response has the `Result=APPROVED` at the end and also has whitespace (or new line).